Refactor CLI into two methods - Run and Parse
This commit separates the responsibility of CLI into two parts:

The cli.Run() method is the primary interface for the struct, returning
the exit code of the command. It will write any errors to the errStream
supplied in the struct.

The cli.Parse() method accepts the list of command line flags and
parses them. It returns a parsed config object, an exit code, and any
errors that occurred.