Pre-parse template functions into objects
This commit pushes the runtime parsing of things like `service` funcs
in the template as early as possible. In a multi-template watcher, it
would be possible to have partial convergence of templates if a 
template later in the slice had an invalid service identifier (for
example). 

The dependencyAcc now parses the strings into objects and returns any
errors that occur.