Diff missing dependencies with unwatched dependencies
As reported in #168, there is a fun data race where Consul Template
loops indefinitely when a template reports missing dependencies but the
watcher is already watching the dependencies (but has not yet returned
data).

This diffs the watcher's list of watches with the dependencies a
template is reporting as missing and only restarts the loop if there
are unwatched dependencies (not just ones that have not returned data
yet).