Add missing locking around Brain's maps.
It's unclear whether a single Brain is intended to be used from multiple
goroutines at once, but Remember takes out the lock when it writes to
these maps, and the new batching goroutine in Runner.Start means it's
possible (if relatively unlikely) for Remember() calls to happen
concurrently with rendering templates.

Take out the locks to be safe.