Introduce a "joiner" function into template functions.
The joiner design is inspired by Jinja2's [joiner](http://jinja.pocoo.org/docs/dev/templates/#joiner).

It allows insertion of a separator into template output in contexts where a loop index is not available
and where the existing, string-only "join" function cannot be used.

An example use case is iterating through Consul's KVs by Key and then joining the underlying values.

    {{ range $key, $pairs := tree "foo" | byKey }}
        $key = {{ $comma := joiner "," }}{{ range $pairs }}{{ sep $comma }}{{ .Value }}{{ end }}
    {{end}}

In this case, no loop index is available (because the tree returns a hash) and string-based join
is not useable because the value types are not strings.