Return a map instead of an array for parseJSON
The default function for parseJSON currently returns an array. This is
problematic.

Fixes GH-635 because it makes using functions like `index` more
difficult due to arrays being unable to have string indexes. This should
be backwards compatible because one can still index a map using a string
(thus treating it as an array).