Add render times to the RenderEvent
This PR brings back render times to the render event. While rebasing
Nomad's integration with the runner I realized there is not enough
information to decide whether to restart a task or even if all templates
have been rendered.

The root of the issue is timing. By the time a caller unblocks on the
render channel and retrieves the map of render events, the event
containing the boolean that it would have rendered or did render could
be replaced by a newer render event. This means that both booleans will
be false, and integrators will not know if the event ever rendered or
rendered again.

By just storing the timestamp, it is always possible to tell if the
template never would have rendered (ts = 0) and if there has been an
update since you have last seen it.