Validate existing template contents
This fixes GH-1000 when a file already exists on disk with the correct
contents, but CT won't exit in -once mode. This is because the template
already exists with the correct content on disk, so "rendered" was
false, but the content "would have rendered", so we have to check that
accordingly.