Fix race condition during file backup.
If the command script moves the file away, and it is timed such that
this happens between `os.Stat()` and `copyFile()`, the `copyFile()`
operation fails and consequently consul-template will terminate.

By having the `copyFile()` operation indicate failure in an ignorable
way, we prevent this race from becoming a termination-cause.

The reason this can happen, is because consul-template does not wait
for the executed command to complete, before any new template rendering
operations take place.