runner: add ForQuiescence to render events
Many render events can potentially be created and processed. They're
stored in a map by template ID so they typically overwrite other
render events. I believe this is all intentional in order to support
the right behaviors for waiting for upstream difference before rendering
a template, but it led to an issue where the `exec` child process
wouldn't ever be created as these render events were being used
to determine completion of rendering.

I believe this current approach could be improved considerably
to be less prone to these types of issues, however this patch
tries to avoid that by marking render events that are generated
from the quiescence internals as such, allowing us to ignore
them later when evaluating if templates have been rendered.