stop tests from hanging/timing out randomly
During dedup leader election, depending on how the go-routines were
ordered, sometimes it would hit a wait/lock condition where it paused.
The default was 15 seconds. This would hang the test for 15 seconds and
sometimes longer. This allows it to be set and sets it during the test
to 100 milliseconds (the shortest amount that worked reliably).

I also made the other timeouts variables instead of constants to keep
them consistent and as a general practice to allow them to be reset
during testing.