fix test issue with token file mtime granularity
When testing the token fetch it writes it twice, making sure it sees the
update. The problem is that if it writes to a fast filesystem (like
tmpfs) it goes faster than the mtime granularity and it doesn't register
the change. So this changes the size of the file as well so it doesn't
rely on the mtime changing.