fix loop to work with different int types
`loop` when passed a scrach or $variable that contained an int as the
function was typed to only take int64. This changes it more like the
other functions that takes a interface{} and type switches/converts it
using reflect. Made it handle integer strings while I was at it.

Fixes #1143