fix error on reading deleted secret
When reading a deleted secret from vault, vault would return meta-data
and no data. The code wasn't setup to handle that and would error out
in the template with an "nil pointer evaluating interface {}" error.

This checks for that and returns the normal "no secret exists" error
that triggers the standard retrying behavior instead of exiting.

Fixes #1198