change backup to use os.Link
Instead of opening a new file and manually copying over the contents,
just use a hard link. This has different tradeoffs vs manually copying
(eg. it isn't supported on plan9), but should generally be more robust
and has no race.

This also changes it to just log an error vs returning it, in the case
of a os.Link returning an error, as that causes consul-template to exit
(which we don't want for this).