alpine docker image doesn't run entrypoint as root
Set USER directrive in the Dockerfile to the consul-template user
created to run consul-template. Also make the UID/GID used for the
consul-template user and group explicit and overrideable. The values
used, if not overridden, are the same as the default values adduser and
addgroup would have used (so no change to actual values).

Made this change as some organizations security policies require it. The
only reason it wasn't done this way before was to run a chown on the
(possibly) bind mounted directories and this is a bad practice anyways
(who knows who the user matches on the external system).

Only change the alpine image this way as the other 2 (scratch and light)
purposely do as little as possible so they don't even create the user.

Fixes #1321