use gnueabihf for 32-bit arm build
The 32 bit arm build that was supposed to work on both arm5 and arm6
won't run on systems without moving files around due to linking/naming
clashes. So we're changing the arm6 build to use
arm-linux-gnueabihf-gcc, which links against the expected libraries.

At this point we're just going to drop support for arm5 as I don't think
anyone is using it and it's starting to become obsolete.

Fixes #1317