Add parseYAML helper
parseYAML works identically to parseJSON but accepts YAML.

YAML is much nicer to use for config files than JSON mainly because
comments can be used.

It should be noted that `parseYAML | toYAML` will strip comments with
the current yaml library, but I don't want to go about changing
that huge of a dependency at this stage.