fix bug looking up versioned vault secrets
The brains/cache is keyed off String() output of a dependency. In this
case the dependency is a versioned vault secret. The problem is that the
vault secret's String() output doesn't contain the version. So the first
version retrieved seeds the cache and all subsequent references use that
version even if they specify a different (or no) version.

Including the version in vault secret String() output fixes this.