move render events to after template command
When using CT as a library, the "standard" way to monitor and know when
to exit is to watch for messages on the TemplateRenderedCh(). The
problem is that the message is sent down that channel before the
commands are run, so the program could exit before the command has run.

The problem is more obvious when CT is run in a transient container,
where even if the commands are run the container might exit before they
finish and they get killed.

Moving the render events to after the command avoids these issues (the
container exit issue also requires a command_timeout > 0s).