remove accidental double call to method
In previous merge/work the renewSecret call was changed to handle it's
returned error but the previous call didn't get removed. It wasn't that
much of an issue as the previous call would have blocked until an error
condition and the second call would have pretty quickly hit the same
issue.