Add Learn links and keep some popular anchors
Learn guides are added for both Consul and Vault to increase awareness
of the different use cases. Many .io sites link to the CT repo markdown
anchors for templating language. This adds the section back with context
to redirect to the new file.

Also moves plugins doc to a new file since it may not be relevant to most
users.