make connect and health service ids different
The IDs need to be different as they are used as the cache keys and
using the same ID means `connect` and `service` would overwrite each
other if both used. (this is bad)

The `connect` function re-uses the health `service` code as it is
identical except for an additional option. When adding connect the ID
was unintentially left the same, it should be different.

Fixes #1458