new docker image build setup
Update to use new standard of downloading application from official
release sources, including cryptographic verification.

Cleans up entrypoint and makes separate init process optional.
Standardizing on a minimal Alpine Linux based image.

Note dumb-init was removed as it is not required to for consul-template,
as it handles all the signaling properly, and docker has it built in so
if a user wants it they can add it by running with `--init`. Shouldn't
normally be necessary but might be useful if you have ill-behaved
template commands.