sync write to file to pass tests on CI systems
On github action's CI setup the file contents weren't always written to
the file when the fetch's monitoring triggers and reads the data (ends
up with empty string). Adding a Sync after the write seems to have fixed
it (it's passed the test in CI multiple times now).

I'm guessing it helps as the CI is running in some kind of VM or
container that has a slow virtual disk sync.