replace shell parsing w/ direct shell call
Instead of trying to parse the shell command line into arguments for
input to exec.Command this replaces that all with passing it to an 'sh
-c' call. This will "fix" the wack-o-mole game of spot fixing issues
that keep coming up with the shell parsing package.

Windows is handled separately (no 'sh') and is setup to only accept/run
a single command.