fix test case file write/stat race
On github actions test file_test fails occasionally when the stat
returns the file has changed before the data has been written. That is
the stat shows an update, it reads the file and can't find the change.

I had attempted to fix this before with a Sync, which seemed to help at
the time, but obviously didn't fix it. This change removes the file Sync
call with a O_SYNC flag on the file open call. Hopefully this will fix
that test (it is getting annoying).