fix connect leaf tests
The service name for the leaf conflicted with another test service name
that was create for other general tests. Connect leaf tests require a
unique/new service name to properly generate the cert (otherwise it
blocks until the cert is generated.. forever in the test case).

This converts the tests to use unique service names.