add new pkiCert template function
The new pkiCert template function and its dependency. Common PKI
dependency testing setup code was placed in vault_common and called in
TestMain. Documentation and template tests round things out.

Most of this is pretty standard new-template-function stuff. The main
differing point for this Dependency is that it is specialized to the
certificate and uses the rendered template destination file as a cache
for the certificate, reading it in and checking the dates to see if it
is still valid. If so it reuses it, it not (or if it's not there) it
fetches a new one. It tries to handle this 'cache' file as forgivingly as
possible, skipping over other CA certificates and text for the PEM
encoded/formatted certificate. It should skip any CA, root or
intermediary/chain.