eliminates confusing use of a slice
The runner keeps a map of template-ids to template-configs for easy
lookup in various places. The template-configs was a slice, though it
could only ever have 1 entry. I'm not sure why the slice is there, maybe
the original author had some idea that never came to fruition or maybe
they just liked iterating... not sure but it I found it very confusing
as IMO using a list/slice heavily implies that there could be more than
one thing in it.

This eliminates that needless slice so that code quits confusing me.