fix name/use of confusing Block flag
The Block flag returned in the ResponseMetadata from Fetch was
confusing. It sounds like it should be about blocking query, but it is
really about a query that would normally block but instead is returning
no data/nil. The logic in this case is to just loop around to call Fetch
again instead of returning the nil data.

The only case of this is KV Get. Consul returns a 404 for KV lookups
on invalid paths but the GO API ignores this error and returns nil data
(and nil error) instead. By having it just loop again it is basically
letting you to have a blocking-like call on a key that doesn't exist yet
(but could come into existance).

This change fixes the naming of the flag to be more obvious in its
meaning, adding a few reminder comments as well as cleaning up a bit
of use from the confusion.