reduce os.Signal noise
User reported missing signals due to a lot of SIGURG signals (Go
scheduler signals). We added code a while back to ignore SIGURG, but
they still are received and take up channel buffer spots. They had to
bump the buffer size to 100 to get it happy.

Reviewing the code there was a handy list of all the signals we accept
in the local signals/ package. So I trimmed out the ignored signals
(SIGURG and SIGCHLD) and passed that to Notify so I wouldn't get the
ignored signals on the channel.

I also bumped the signal channel size to 10 to be a bit more flexible
there as well.