pki certs should renew when expired
The code handled the case of getting a cert and using the version on the
disk as the cache but missed looping over that logic when the one on
disk had expired.

This adds the looping needed to renew the cert when expired and extends
the refetching test to validate this functionality.