pkiCert now gets the Cert, Private key and CA (#1591)
*pkiCert now gets the Cert, Private key and CA

Reworked `PKICert` to return an object with the Cert, the CA and the
Private Key. They are stored as stings on a returned struct with
appropriate names (Cert, Key, CA).

Made it so that the `Data` intermediary variable works like secret so
the differing behavior between `secret` and `pkiCert` wouldn't be
a snag.

Includes updated tests and documentation.