skip reload signal if set to "" or signull
Converted odd SIGNIL type that wouldn't work with syscall type assertion
and allow for differentiating between it and an error into SIGNULL,
`syscall.Signal(0)`. SIGNULL works here as it allows for the type
assertion to work and for a test to skip the signal when found. It is
safe as it already exists in practice in a similar role with the POSIX
`kill` command where it indidates no signal.

The other changes here are to replace reference to the hold SIGNIL.