handle wrap_ttl tokens in VaultAgentTokenQuery
The VaultAgentTokenQuery dependency is used to monitor the vault-agent
token file (when present) and update the token as the file gets updated.
It didn't reuse any code from client_set so the change there didn't fix
this.

This adds a call to the function to pull the token from the json
structure to the dependency code.