vault token management moved into separate watcher
Refactor the vault token management (renewing, tokenfile watching,
unwrapping, etc) into a dedicated watcher that is only responsible for
that.

Done to encapsulate the vault management token into one place where it
can be more easily understood and tested. It was scattered about and
inconsistent (eg. it only tried to unwrap the first token).

Strips the vault token code out of current watcher and client_set.

Most of this commit is the new test suite for this code as it wasn't
really tested before.