isChownNeeded bad when only user/group (!both) set
isChownNeeded returned the wrong result if either (exclusive) user or
group were set but not the other. It uses a guard value but only
compares both at the start and doesn't take this into account on the
final comparison which compares both against the current values.

IE. it always returns true in these cases as the guard value is never
going to be the same as an real file value.