use internal signal implementation for force kill
Force killing was using the builtin Process.Kill() command that signals
the process with os.Kill but doesn't take into account if Setpgid has
been set so doesn't kill any subprocesses (if the command was run as
a shell string that uses Setpgid to be sure signals propogate to any
spawned sub-processes).

This changes it to use the internal signal method which checks Setpgid
and uses the process group id to kill all processes in the group.