config default tweaks based on cts research
See this ticket for some details...
https://github.com/hashicorp/consul-terraform-sync/pull/164

Summary: http.MaxIdleConnsPerHost was default to a very low value that
caused TCP connections to not be reused across dependency monitoring.
Only ~1 connections were being reused from the initial set of hcat
requests to Consul, and the remaining dependencies required establishing
a new connection.