Make setpgid configurable for exec commands
#1494 added the setpgid setting to ensure signals are propagated to all
processes and not just the subshell. However, this might not always be
desirable.

For instance, servers like uWSGI run workers in subprocesses and wait
for workers to finish processing active requests for a while before
restarting them to support graceful reloading. consul-template sending
the reload signal to all processes, sends singals to these worker
processes in uWSGI as well which can interfere with graceful reloads.

Currently, we are working around this by using the array form of the
command which does not set the pgid on the child process. But I feel
it might be useful to make this option configurable as well.