writeToFile: allow empty permission to preserve the existing
Allow empty permission to preserve the existing perm if the file exists.
Perm 0664 will be used for the created files if it's not specified.

This can be useful for the cases where the file already exists with the
right perms but the vault agent process doesn't have access to chmod it
but only read/write to it.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>