Hack week project to teach consul-template how to retrieve secrets from HVS.
Currently oriented around HVS in a PRDE, since that's convenient for testing.  As such, hacked the tests to put test instances of Nomad and Vault on 127.0.0.2 instead of 127.0.0.1, as they would otherwise collide with PRDE proxy.  To run the tests it's therefore necessary to run `sudo ifconfig lo0 alias 127.0.0.2` on MacOS or `sudo route add -host 127.0.0.2 dev lo` on Linux.

The following env vars also need to be set:
* HCP_API_TLS=disabled
* HCP_API_ADDRESS=localhost:28081
* PRDE_SERVICE_NAME=ncabatoff01-9yir6h5q (use the part between "registry" and ".dev.pedp" in the "docker registry" output line of hcloud hcp dev proxy)
* HCP_CLIENT_SECRET
* HCP_CLIENT_ID
* HCP_PROJECT_ID
* HCP_ORGANIZATION_ID